% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getterm.gene.R
\name{getterm.gene}
\alias{getterm.gene}
\title{get all the term genes under the term "biological process", "molecular function" or "cellular composition"}
\usage{
getterm.gene(
  term,
  gene2go_human,
  graph_path,
  relation = c("biological_process", "molecular_function", "cellular_component")
)
}
\arguments{
\item{term}{a matrix or data.frame contains the informations of the terms in the Gene Ontology database. Download from http://archive.geneontology.org/lite/2019-07-06/. By default the variables are taken from environment(formula).}

\item{gene2go_human}{a matrix or data.frame contains the informations of the geneID matching to the GOID. Download from https://ftp.ncbi.nih.gov/gene/DATA/ and should be preprocessed to get the matrix or data.frame of geneID matching to GOID of human beings. By default the variables are taken from environment(formula).}

\item{graph_path}{a matrix or data.frame contains the informations of relations of parent term and son term. Doownload from http://archive.geneontology.org/lite/2019-07-06/. By default the variables are taken from environment(formula).}

\item{relation}{a character string specifying the research direction , must be one of "biological_process","molecular_function" or "cellular_component"}
}
\value{
a matrix whose row is geneID and col is termID
}
\description{
the function can get all the offspring pathway and their entire genes under "biological_process" , "molecular_function" or "cellular_component"
}
\examples{
term_gene<-getterm.gene<-function(term,gene2go_human,graph_path,relation="biological_process")
}
