% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.redundance.R
\name{remove.redundance}
\alias{remove.redundance}
\title{remove redundant pathway}
\usage{
remove.redundance(
  stable.pair,
  reverse.pair,
  gene.minimum = 1,
  patients,
  re_graph,
  holidyregpath,
  term_gene_list,
  threads = 1L
)
}
\arguments{
\item{stable.pair}{a matrix or data.frame of stable pairs from normal samples with two columns, the expression of the genes in the first column is higher than that in the second column.}

\item{reverse.pair}{a matrix or data.frame of reverse pairs from disease samples with two columns, the expression of the genes in the first column is higher than that in the second column.}

\item{gene.minimum}{an integer value to control the minimum amount of the genes of the terms. The deault is 1.}

\item{patients}{a matrix of data.frame of tumor samples , the first column must be the geneID ,and tumor samples start with the second column.}

\item{re_graph}{a matrix or data.frame with three columns about the relationship between parent term and whose son term. The first to third column is parent termID, son termID and distance between parent term and son term separately.}

\item{holidyregpath}{a matrix or data.frame with three columns depicts the coorelation levels of terms and disease. Which computed by method get.holidyregpath. The first to third column is termID, p-value and FDR value.}

\item{term_gene_list}{a list of term genes. list type. each list is a vector type, containing all the genes of a term.By default the variables are "BPtermgene_list" and taken from environment(formula).}

\item{threads}{an integer value to make sure how many threads would be used to complete the computation.}
}
\value{
a list.The first list is all terms and their redundant parent terms. of all the nonredundant holistic-dyregulated pathway
}
\description{
this function is used to remove redundant parent or son terms from all of global dysregulation terms
this function support parallel computation . You need to set thread numbers to make sure how many threads do you need to use.
}
\examples{
stable.pair<-t(combn(sample(1:10,10),2));
geneid<-1:10;
samples<-runif(100,min = 0,max = 50);
patients<-matrix(c(geneid,samples),nrow = 10,byrow=F);
reverse_pairs<-revpairs(stable.pair,patients,threshold=0.05,spairs_threshold=0.99,threads=1L,capacity=300000L)
global_dysterm<-get.globdysterm(stable.pair=stable.pair,reverse.pair=reverse_pairs,threads = 1,gene.minimum = 2,term_gene_list = BPtermgene_list,patients = patients,threshold = 0.05)
nonredundance<-remove.redundance(stable.pair = stable.pair,reverse.pair = reverse_pairs,gene.minimum = 2,patients = patients,re_graph = bp_graph,holidyregpath = global_dysterm,term_gene_list = BPtermgene_list,threads = 1)
}
