% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revpairs.R
\name{revpairs}
\alias{revpairs}
\title{get reverse pairs from all the tumor samples}
\usage{
revpairs(
  stable.pair,
  patients,
  threshold = 0.05,
  spairs_threshold = 0.99,
  threads = 1L,
  capacity = 300000L
)
}
\arguments{
\item{stable.pair}{a matrix or data.frame of stable pairs from normal samples with two columns, the expression of the genes in the first column is higher than that in the second column.}

\item{patients}{a matrix of data.frame of tumor samples , the first column must be the geneID ,and tumor samples start with the second column.}

\item{threshold}{a numeric value which is used to control false discovery rate under the p_value of the chip-square test or fisher's exact probability test , default is 0.05.}

\item{spairs_threshold}{a threshold same with the "threshold" in function "spairs".}

\item{threads}{an integer value to make sure how many threads you will use to complete the computation}

\item{capacity}{an integer value to depict the computation capacity, ruling how many lines of stable pairs would be computed within one time.the default is 300000}
}
\value{
a matrix containing four columns respectively represent higher expression gene , lower expression gene , p_value under binomial distribution , false discovery rate under p.adjust
}
\description{
the function is used to get reverse gene pairs from a host of disease samples.
this function support parallel computation . You need to set thread numbers to make sure how many threads do you need to use .
}
\examples{
stable.pair<-t(combn(sample(1:10,10),2));
geneid<-1:10;
samples<-runif(100,min = 0,max = 50);
patients<-matrix(c(geneid,samples),nrow = 10,byrow=F);
reverse_pairs<-revpairs(stable.pair,patients,threshold=0.05,spairs_threshold=0.99,threads=1L,capacity=300000L)
#compute with parallel
reverse_pairs<-revpairs(stable.pair,patients,threshold=0.05,spairs_threshold=0.99,threads=10L,capacity=300000L)
}
