% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spairs.R
\name{spairs}
\alias{spairs}
\title{get stable pairs from normal samples}
\usage{
spairs(data1, gid1, threshold = 0.99)
}
\arguments{
\item{data1}{a (normal sample) dataset which is used to screen the stable_pair.Notable that rownames is gene identifier(geneID/genesymbol/ensembleID).}

\item{gid1}{a vector of gene indentifier which can completely match data1}

\item{threshold}{a threshold value is used to stipulate that a gene pair maintain the circumstance of > the threshold so that the gene pair should be retained.The threshold should between 0-1, and the default is 0.99.}
}
\value{
a list which contains a dataframe of stable pair and a runtime
}
\description{
the function is used to get stable gene pairs from a host of normal samples.
}
\examples{
geneid<-1:10;
samples<-runif(100,min = 0,max = 50);
data1<-matrix(c(geneid,samples),nrow = 10,byrow=F);
stable_pairs<-spairs(data1 = data1,gid1=geneid,threshold = 0.99)

}
